Tạo được đơn hàng nhưng chưa confirmed được

+++ shop/backend/src/main/java/com/g5/techdevices/techstore/dtos/cart/AddToCartRequestDTO.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 // Dùng cho request thêm vào giỏ
 public class AddToCartRequestDTO {
-    private Integer variantId;
+    private Long variantId;
     private int quantity;
 }
